refactor: extract MobileBottomSheets from MobileLayout
Extract the three mobile bottom sheets (recycle bin, tag filter, daily
summary with progress ring and stat rows) into a MobileBottomSheets
component using a Fragment wrapper.

MobileLayout drops from ~2619 to ~2194 lines.
